Serialization fix, cleanup.

+++ user.api/Common/Models/DTO/Base/BaseDTO.cs
@@ -3,6 +3,6 @@
 public abstract class BaseDTO : IDTO
 {
     public Guid? Id { get; set; }
-    public long? CreatedOn { get; set; }
+    public int? CreatedOn { get; set; }
 }
