+++ user.api/Common/Models/DTO/Base/IDTO.cs
@@ -3,6 +3,6 @@
 public interface IDTO
 {
     Guid? Id { get; set; }
-    long? CreatedOn { get; set; }
+    int? CreatedOn { get; set; }
 }
