+++ user.api/Common/Models/Data/Base/BaseEntity.cs
@@ -3,6 +3,6 @@
 public abstract class BaseEntity : IEntity
 {
     public Guid id { get; set; }
-    public long created_on { get; set; }
+    public int created_on { get; set; }
 }
